commit 6

+++ src/main/java/com/safa/modasluz/modelos/Producto.java
@@ -31,5 +31,5 @@ public class Producto {
     private Color color;
 
     @Column(name = "precio")
-    private Integer precio;
+    private Double precio;
 }
